+++ src/frontend/apps/e2e/__tests__/app-impress/doc-export.spec.ts
@@ -30,7 +30,7 @@ test.describe('Doc Export', () => {
     await expect(page.getByTestId('modal-export-title')).toBeVisible();
     await expect(
       page.getByText(
-        'Export your document to print or download in .docx, .odt, .pdf or .html(zip) format.',
+        'Export your document to download in .docx, .odt, .pdf or .html(zip) format.',
       ),
     ).toBeVisible();
     await expect(page.getByRole('combobox', { name: 'Format' })).toBeVisible();
@@ -303,12 +303,7 @@ test.describe('Doc Export', () => {
 
     await overrideDocContent({ page, browserName });
 
-    await clickInEditorMenu(page, 'Download');
-
-    await page.getByRole('combobox', { name: 'Format' }).click();
-    await page.getByRole('option', { name: 'Print' }).click();
-
-    await page.getByRole('button', { name: 'Print' }).click();
+    await clickInEditorMenu(page, 'Print');
 
     await expect(page.locator('#print-only-content-styles')).toBeAttached();